+++ azext_iot/central/models/v1_1_preview/query_response.py
@@ -1,0 +1,10 @@
+# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+
+class QueryResponse:
+    def __init__(self, response: dict):
+        self.results = response.get("results")