+++ azext_iot/common/utility.py
@@ -462,7 +462,9 @@ def ensure_iotdps_sdk_min_version(min_ver):
     try:
         from azure.mgmt.iothubprovisioningservices import __version__ as iot_sdk_version
     except ImportError:
-        from azure.mgmt.iothubprovisioningservices._version import VERSION as iot_sdk_version
+        from azure.mgmt.iothubprovisioningservices._version import (
+            VERSION as iot_sdk_version,
+        )
 
     return version.parse(iot_sdk_version) >= version.parse(min_ver)