+++ HISTORY.rst
@@ -6,6 +6,14 @@ Release History
 0.12.0
 +++++++++++++++
 
+**IoT Central updates**
+
+* Adds new preview commands (v1.1-preview)
+
+  - Query (az iot central query)
+  - Destination (az iot central export destination)
+  - Export (az iot central export)
+
 ** General Updates **
 
 * The IoT extension officially supports Python 3.10.